Break up http-related trace code; Fix tests and refs
