+++ trace/es/elasticsearch.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 
 	"github.com/beatlabs/patron/trace"
+	trace_http "github.com/beatlabs/patron/trace/http"
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/elastic/go-elasticsearch/v8/esapi"
 	"github.com/elastic/go-elasticsearch/v8/estransport"
@@ -58,7 +59,7 @@ func endSpan(sp opentracing.Span, rsp *http.Response) {
 	// In cases where more than one host is given, the selected one is only known at this time
 	sp.SetTag(respondentTag, rsp.Request.URL.Host)
 
-	trace.FinishHTTPSpan(sp, rsp.StatusCode)
+	trace_http.FinishSpan(sp, rsp.StatusCode)
 }
 
 type transportClient struct {
