+++ trace/http/http.go
@@ -6,11 +6,19 @@ import (
 	"time"
 
 	"github.com/beatlabs/patron/correlation"
+	"github.com/beatlabs/patron/log"
 	"github.com/beatlabs/patron/reliability/circuitbreaker"
 	"github.com/beatlabs/patron/trace"
 	"github.com/opentracing-contrib/go-stdlib/nethttp"
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/ext"
+)
+
+const (
+	// HTTPComponent definition.
+	HTTPComponent = "http"
+	// HTTPClientComponent definition.
+	HTTPClientComponent = "http-client"
 )
 
 // Client interface of a HTTP client.
@@ -48,8 +56,8 @@ func New(oo ...OptionFunc) (*TracedClient, error) {
 func (tc *TracedClient) Do(ctx context.Context, req *http.Request) (*http.Response, error) {
 	req = req.WithContext(ctx)
 	req, ht := nethttp.TraceRequest(opentracing.GlobalTracer(), req,
-		nethttp.OperationName(trace.HTTPOpName(req.Method, req.URL.String())),
-		nethttp.ComponentName(trace.HTTPClientComponent))
+		nethttp.OperationName(OpName(req.Method, req.URL.String())),
+		nethttp.ComponentName(HTTPClientComponent))
 	defer ht.Finish()
 
 	req.Header.Set(correlation.HeaderID, correlation.IDFromContext(ctx))
@@ -79,4 +87,31 @@ func (tc *TracedClient) do(req *http.Request) (*http.Response, error) {
 	}
 
 	return r.(*http.Response), nil
+}
+
+// Span starts a new HTTP span.
+func Span(path, corID string, r *http.Request) (opentracing.Span, *http.Request) {
+	ctx, err := opentracing.GlobalTracer().Extract(opentracing.HTTPHeaders, opentracing.HTTPHeadersCarrier(r.Header))
+	if err != nil && err != opentracing.ErrSpanContextNotFound {
+		log.Errorf("failed to extract HTTP span: %v", err)
+	}
+	sp := opentracing.StartSpan(OpName(r.Method, path), ext.RPCServerOption(ctx))
+	ext.HTTPMethod.Set(sp, r.Method)
+	ext.HTTPUrl.Set(sp, r.URL.String())
+	ext.Component.Set(sp, "http")
+	sp.SetTag(trace.VersionTag, trace.Version)
+	sp.SetTag(correlation.ID, corID)
+	return sp, r.WithContext(opentracing.ContextWithSpan(r.Context(), sp))
+}
+
+// FinishSpan finishes a HTTP span by providing a HTTP status code.
+func FinishSpan(sp opentracing.Span, code int) {
+	ext.HTTPStatusCode.Set(sp, uint16(code))
+	ext.Error.Set(sp, code >= http.StatusInternalServerError)
+	sp.Finish()
+}
+
+// OpName return a string representation of the HTTP request operation.
+func OpName(method, path string) string {
+	return method + " " + path
 }
